Custom contract name: LogColor
Renamed GreenConsoleLogger to ColoredConsoleLogger. Now it expects a color
parameter in its constructor. The constructor is marked with ImportingConstructor
attribute and the color parameter is marked with `[Import("LogColor")]`.

Here `LogColor` is a name of a contract. I've added a property
`ConsoleLogger.DefaultContract` that exports this contract. As a result, the
value of this property will be used a parameter value.

Note that I had to make logger plugins lazy. Otherwise MEF won't be able
to compose parts because of a dependency loop.